requirements.txt
# Core dependencies for Agentic Edu-RAG System
# LLM and Agent Framework
openai>=1.55.0
langgraph>=0.2.74
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Vector Database and RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.5.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Evaluation and Metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For development server
uvicorn>=0.23.0
fastapi>=0.100.0
